requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.3
amqp==5.1.1
annotated-types==0.5.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
bcrypt==4.0.1
bidict==0.22.1
billiard==3.6.4.0
blinker==1.6.1
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.33.0
colorama==0.4.6
cryptography==40.0.2
dataclasses-json==0.5.14
DateTime==5.1
dnspython==2.3.0
email-validator==1.3.1
eventlet==0.33.3
Flask==2.2.3
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SocketIO==5.3.3
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
frozenlist==1.3.3
google-auth==2.17.3
google-auth-oauthlib==1.0.0
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kombu==5.2.4
langchain==0.0.268
langsmith==0.0.25
loguru==0.7.0
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
oauthlib==3.2.2
openai==0.27.5
packaging==23.1
passlib==1.7.4
pinecone-client==2.2.2
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
PyJWT==2.6.0
pyOpenSSL==19.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.4.1
python-http-client==3.3.7
python-socketio==5.8.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
sendgrid==6.10.0
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.9
SQLAlchemy-serializer==1.4.1
starkbank-ecdsa==2.2.0
stripe==5.4.0
tenacity==8.2.3
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==1.26.15
uuid==1.30
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
win32-setctime==1.1.0
WTForms==3.0.1
yarl==1.9.2
zope.interface==6.0
